chatbtc-telegram/requirements.txt
openai
python-telegram-bot==13.5
lnd-grpc-client
python-dotenv
qrcode
requests
telegram